third_party/WebKit/LayoutTests/fast/events/wheel/wheel-scroll-latching-on-scrollbar.html

<!DOCTYPE HTML>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>

  body {
    margin: 0px;
    height: 100px;
    width: 100px;
  }
  #parentDiv {
    background-color: #FF7F7F;
    height: 100px;
    width: 100px;
    overflow: scroll;
  }
  #content1 {
    height: 120px;
    width: 120px;
  }
  #childDiv {
    background-color: #84BE6A;
    height: 80px;
    width: 80px;
    overflow: scroll;
  }
  #content2 {
    height: 100px;
    width: 100px;
  }
</style>

<div id="parentDiv">
  <div id="content1">
    <div id="childDiv">
      <div id="content2">
      </div>
    </div>
  </div>
</div>

<script>
var childDiv = document.getElementById('childDiv');
var parentDiv = document.getElementById('parentDiv');
var rect = childDiv.getBoundingClientRect();
const GESTURE_SOURCE_TYPE = 2; // MOUSE_INPUT from synthetic_gesture_params.h

function scrollDown(pixels_to_scroll, start_x, start_y, speed_in_pixels_s) {
  return new Promise((resolve, reject) => {
    chrome.gpuBenchmarking.smoothScrollBy(pixels_to_scroll,
                                          resolve,
                                          start_x,
                                          start_y,
                                          GESTURE_SOURCE_TYPE,
                                          'down',
                                          speed_in_pixels_s);
  });
}

const MAX_RAF = 1000;
function waitForAnimationEnd() {
  var last_child_scroll_offset = childDiv.scrollTop;
  var last_parent_scroll_offset = parentDiv.scrollTop;
  var last_changed_count = 0;
  return new Promise((resolve, reject) => {
    function tick(raf_count) {
    // We requestAnimationFrame either for 1000 frames or until 20 frames with
    // no change have been observed.
      if (raf_count >= MAX_RAF || raf_count - last_changed_count > 20) {
        resolve();
      } else {
        if (childDiv.scrollTop != last_child_scroll_offset ||
            parentDiv.scrollTop != last_parent_scroll_offset) {
          last_changed_count = raf_count;
          last_child_scroll_offset = childDiv.scrollTop;
          last_parent_scroll_offset = parentDiv.scrollTop;
        }
        requestAnimationFrame(tick.bind(this, raf_count + 1));
      }
    }
    tick(0);
  });
}

promise_test(t => {
  return scrollDown(1000, rect.right - 5, rect.bottom - 5, 4000)
  .then(waitForAnimationEnd)
  .then(() => {
    assert_equals(childDiv.scrollTop, childDiv.scrollHeight - childDiv.clientHeight,
        "childDiv must be fully scrolled");
    assert_equals(parentDiv.scrollTop, 0, "parentDiv shouldn't scroll at all");
  });
}, "Scrolling on scrollbar of the child div doesn't propagate to the parent.");

</script>
